requirements.txt
openai>=1.0.0,<2.0.0
azure-identity>=1.16.0
python-dotenv>=1.0.1
ragas
deepeval
langchain-openai>=0.1.0